// frontend/lib/api/models/UserResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserResponse = {
    id: number;
    username: string;
    email: string;
    full_name: (string | null);
    bio: (string | null);
    profile_image: (string | null);
    followers_count?: number;
    following_count?: number;
    is_following?: boolean;
    created_at: string;
};